voice-agent/pyproject.toml
[project]
name = "livekit-voice-agent"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "livekit-agents[silero]>=1.5.2",
    "livekit-plugins-noise-cancellation>=0.2.5",
    "python-dotenv>=1.2.2",
]
